Changes to version switching functionality * Getting rid of fallback concept. Moving default controllers to v1. * Throw exception in version switcher if controller not found * Simplifying controller resolving in version switcher
