+++ tests/phpunit/Unit/VersionSwitcherTest.php
@@ -4,8 +4,10 @@ namespace Tests\Unit;
 
 use Tests\TestCase;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Config;
 use App\Http\Middleware\VersionSwitcher;
 use App\Exceptions\AcceptHeaderWrongFormatException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class VersionSwitcherTest extends TestCase
 {
@@ -30,6 +32,8 @@ class VersionSwitcherTest extends TestCase
      */
     public function testThatMiddlewareSwapsControllerPathDependingOnVersion(string $version, string $expectedController): void
     {
+        Config::set('api.version', '1');
+
         $request = $this->getMockBuilder(Request::class)
             ->onlyMethods(['route'])
             ->getMock();
@@ -39,8 +43,13 @@ class VersionSwitcherTest extends TestCase
             ->method('route')
             ->will($this->returnValue($this->configuredRoute));
 
+        // The v1000 controller does not exist. We expect a not found exception to be thrown in that case.
+        if ($version == '1000') {
+            $this->expectException(NotFoundHttpException::class);
+        }
         $middleware = new VersionSwitcher();
-        $middleware->handle($request, function ($request) use ($expectedController) {
+
+        $middleware->handle($request, function ($request) use ($version, $expectedController) {
             $this->assertSame($expectedController, app()->getAlias($this->baseController));
         });
     }
@@ -65,13 +74,13 @@ class VersionSwitcherTest extends TestCase
     {
         return [
             // If we are using version 1 we should get the configured controller
-            // because there is no list controller present in the v1 directory.
-            ['1', 'App\Http\Controllers\ListController'],
+            // // because there is no list controller present in the v1 directory.
+            ['1', 'App\Http\Controllers\v1\ListController'],
             // If we use version 2 we should get the v2 list controller.
             ['2', 'App\Http\Controllers\v2\ListController'],
-            // If we are using version 3 we should get the configured controller
-            // because there is no list controller present in the v1 directory.
-            ['3', 'App\Http\Controllers\ListController'],
+            // If we are using version 1000 we should get an exception
+            // because there is no list controller present in the v1000 namesapce.
+            ['1000', ''],
         ];
     }
 }
